mlaas-demos/face_detection_api/pyproject.toml
[tool.poetry]
name = "face-detection-api"
version = "0.1.0"
description = ""
authors = ["Henry Ruiz  <henryruiz22@gmail.com>"]
readme = "README.md"
packages = [{include = "face_detection_api"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
pillow = "^9.3.0"
matplotlib = "^3.6.2"
google-cloud-vision = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
